Created the function unblockNewVotingToStart(0).
There are several folks trying to start the voting, but they are
blocked when the voting status is set as over, however when the
extension option wins, the vote status is not set as over to allow a
new voting when the extension time to expires.

To fix the voting starting right again after the map extension,
blocking it by 2 minutes should be big enough to not block the new
voting after the map extension time expires.