Allowed the map_populateList(5) not load duplicated lines:
Currently the empty cycle behavior is based on the function
map_getNext(4). Therefore this is a problem because it will create
infinity empty cycle loops between duplicated maps on the empty map
cycle.

So here I have to options:
a) change all the empty cycle algorithm
b) do not load duplicated maps.

As anyone until now complained about this behavior, I choose the
former as easier to implement.