Fixed the `runoff` voting starting twice due the vote round start
feature. This happened because originally the plugin uses the same
function to start the runoff voting and also the normal voting,
moreover depending purely on global state to correctly chooses
whether it is a normal voting or a runoff voting.

Now I should fixed it splitting the vote start function, with one
for runoff and other for the normal voting.