Until know there are implemented 926 Unit Tests succeed,
the Unit Test count may vary due the Unit Test
test_handleServerStart(0) creates a variable number of tests
accordingly to the constant MAX_SERVER_RESTART_ACCEPTABLE.

Therefore, every time that constant is changed the number of
tests is going to vary. Moreover it is important to log all the
unit tests count change as on the title of the commit message,
for easy monitoring/control.